say-home-ai/DbHelper.py
import mysql.connector
import os

class DbHelper:
    @staticmethod
    def get_db():
        return mysql.connector.connect(
            host=os.getenv("DB_HOST", "localhost"),
            port=int(os.getenv("DB_PORT", 3306)),
            user=os.getenv("DB_USER", "root"),
            password=os.getenv("DB_PASSWORD", "root"),
            database=os.getenv("DB_NAME", "say_home_db")
            )