say-home-ai/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
dotenv==0.9.9
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
idna==3.11
mysql-connector==2.2.9
mysql-connector-python==9.6.0
pydantic==2.13.0
pydantic_core==2.46.0
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
